protect endpoints through user auhorization

+++ src/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from src.schemas import PostCreate, PostResponse
-from src.db import Post, create_database_tables, get_async_session
+from src.db import Post, User, create_database_tables, get_async_session
 from src.album import imagekit
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
@@ -35,6 +35,7 @@ app.include_router(fastapi_users.get_users_router(UserRead, UserCreate), prefix=
 async def upload_file(
 		file: UploadFile = File(...),
 		caption: str = Form(""),
+		user: User = Depends(current_user),
 		session: AsyncSession = Depends(get_async_session)
 ):
 	"""Create a temporary copy of the uploaded file"""
@@ -64,6 +65,7 @@ async def upload_file(
 		)
 
 		post = Post(
+			user_id = user.id,
 			caption=caption,
 			url=upload_output.url,
 			file_type="video" if file.content_type.startswith("video/") else "image",
@@ -83,10 +85,16 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-		session: AsyncSession = Depends(get_async_session)
+		session: AsyncSession = Depends(get_async_session),
+		user: User = Depends(current_user)
 ):
-	result = await session.execute(select(Post).order_by(Post.created_at.desc()))
-	posts = [row[0] for row in result.all()]
+	output = await session.execute(select(Post).order_by(Post.created_at.desc()))
+	posts = [row[0] for row in output.all()]
+
+	output = await session.execute(select(User))
+	users = [row[0] for row in result.all()]
+	user_dict = {user.id: user.email for user in users}
+
 
 	posts_data = []
 
@@ -94,11 +102,14 @@ async def get_feed(
 		posts_data.append(
 			{
 			"id": str(post.id),
+			"user_id": str(post.user_id),
 			"caption": post.caption,
 			"url": post.url,
 			"file_type": post.file_type,
 			"file_name": post.file_name,
-			"created_at": post.created_at.isoformat()
+			"created_at": post.created_at.isoformat(),
+			"is_owner": post.user_id == user.id,
+			"email": user_dict.get(post.user_id, "user not found!")
 		}
 	)
 
@@ -106,7 +117,7 @@ async def get_feed(
 
 """Deleting a post"""
 @app.delete("/posts/{post_id}")
-async def delete_data(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_data(post_id: str, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_user)):
 	try:
 		post_uuid = uuid.UUID(post_id)
 
@@ -115,6 +126,9 @@ async def delete_data(post_id: str, session: AsyncSession = Depends(get_async_se
 
 		if not post:
 			raise HTTPException(status_code=404, detail="Post does not exists!")
+
+		if post.user_id != user.id:
+			raise HTTPException(status_code=403, details="Unauthorized action, login with your account to delete post!")
 
 		await session.delete(post)
 		await session.commit()
